pyproject.toml
[project]
name = "santander-desafio"
version = "0.1.0"
description = "API of santander challenge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.118.0",
    "google-genai>=1.41.0",
    "graphviz>=0.21",
    "neo4j>=6.0.2",
    "networkx>=3.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.11.0",
    "pygraphviz>=1.14",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "tenacity>=9.1.2",
]
